Add a modal dialog when a DOM scrape is in progress
This is currently only used when scraping the list of playlists
(which currently cannot be triggered by the user).
Upon completion, a 'copy to clipboard' button is added to the dialog,
which copies the list to the clipboard and closes the dialog.

Removed the 'Playlist' class that wasn't really necessary.
Added a 'ScrapeInProgressDialog' class to handle the new dialog.
Reduced the scrape timeout to 2.5 seconds down from 4 seconds.