Support scraping playlist names & show dialog during scrapes (#19)
New Functionality:
Added support for scraping a list of all playlist names from the DOM in 
the YouTube Music 'library/playlists' page. This can be triggered from 
a new context menu item accessible only on that page. Once the scrape 
is complete, the user is prompted with the option of copying the 
resulting list of playlist names to the clipboard.

UI:
Added a dialog modal that is displayed whenever a scrape is in progress.
In the case of a playlist names scrape, the dialog prompts the user to 
copy the resulting list to the clipboard once the scrape is complete. 

Scraping Logic:
Replaced track scraping logic with a more abstract function that can 
scrape elements in a list given a set of criteria. 
This function is now used to scrape a list of playlists from the DOM. 
This function returns a promise, allowing for async/await usage in the 
functions that call it, which makes the related message passing clearer.
Reduced the scrape timeout to 2.5 seconds down from 4 seconds.
Added new classes to support the new 'scrape in progress' dialog modal.
Added a helper function to convert an array to a single-column CSV.

Miscellaneous:
Now only attempt to display a tracktable for the 'stored' tracklist if 
it exists, rather than error out.