Table names for deltas copied to clipboard are now in 2nd column
Before, each table name in the delta tables copied to clipboard would
have the table name appear in the first column - the same one used for
the index value. This made it slightly less convenient to keep the
index column width small; it required some simple merging. Now, the
first/index column in the header row is kept blank, and the table is in
the second column, which is the one where track names appear,
which require a longer column width anyway. This means the table can be
set to "fit width to content" right after import/paste in one go without
more specific changes required.