Improve track table UI & address misc. edge cases (#21)
UI Improvements - Track Table Layout:

A blank string is now used for a cell's value if the metadata value is 
undefined. This ensures that all values show up in the correct columns 
of a track table in the extension's UI.

Track tables now have a 'fixed' layout, so all the columns always fit 
within the popup window. Added new classes for some of the track table 
header cells, including hardcoded width percentages for index, 
duration, and unplayable columns.

UI Improvements - Exported/Copied Delta Track Table Names:

Track tables copied to clipboard now list their name in the 2nd column. 
(Examples of table names are 'Added', 'Removed', etc.).
Before, each table in the delta tables copied to clipboard would have 
its name appear in the first column - the same one used for the index 
value. This made it slightly less convenient to keep the index column 
width small. (It required some merging after being pasted into a 
spreadsheet). Now, the first/index column in the header row is kept 
blank, and the table name is included in the second column, which is 
the one where track names appear, and requires a longer column width 
anyway. This means the table can be set to "fit width to content" right 
after import/paste in one go without more specific changes required.

Also added support to include an overall table name (e.g. using the 
tracklist title) in the first row of the delta track table CSV. 

Miscellaneous Bug Fixes & Improvements:

Added better handling of case where 'duration' isn't a string. Now, if 
the duration value isn't a string in both the scraped and stored data 
(e.g. undefined in both), this won't result in the tracks not matching. 

Added a mapping function for tracklist titles. This allows GPM data to 
be used for deltas even if the tracklist titles don't exactly match. 
This is helpful for cases where the tracklist still exists but its 
title may have been changed.

Context menus are now created on installation of the extension. Prior 
to this, they were created on firebase initialization, which could 
happen more than once and would result in an error.

Default preferences are now set on installation of the extension. Prior 
to this, on initial installation, no 'comparison method' preference 
would be set in storage, which could lead to some UI and logic issues.

Added some error handling and messaging in the case where GPM
library data can't be found in local storage. Now clear warnings are
printed but there are no errors. The extension icon will fallback
to a question mark.

Updated error throwing in various places to follow a consistent format.