Add fnc to generate list of uploaded GPM tracks
There isn't a list of specifically just uploaded tracks in GPM storage.
This fnc starts by taking all tracks in the 'ALL MUSIC' GPM tracklist.
Then, each of these tracks that isn't also found in 'Added from Sub.'
gets added to the 'Uploaded' list.
In the future, it may be best to just store this list in local storage.

Also updated the fnc to add playlist data to tracks so it takes a
list of playlist data objects a list of playlist titles to exlude. This
makes it easy to use for both YTM and GPM.