Add helper fnc to filter tracks & apply tracklist mapping
The new function takes an initial tracklist title, as well as an
optional list of titles of tracklists to filter out. It filters out any
tracks contained within these tracklists, and then adds a tracklist
mapping to each of the remaining/filtered tracks. The tracklist mapping
is just a comma-separated string of all tracklists in which that track
appears, which can be used to identify issues.

The helper function to filter a tracklist now supports as a parameter a
list of tracklists to filter out, instead of custom 'filter' objects
which were used previously and a bit confusing/overcomplicated.