Separate logic to filter out a single tracklist
The filterTracklist function has been split into two:
- filterOutTracklist
- filterOutTracklists
The former handles the majority of the logic needed to filter a single
tracklist ouf of an original array of tracks. The latter handles
applying this against multiple tracklists and continuously paring down
the original array of tracks. This split should make the logic easier
to follow, and also allows the former function to be used elsewhere.

Also, neither of these functions are exported outside of the module now.