Fix a couple bugs relating to CSV generation
Fixed a bug where the last element in a generated csv row string didn't
have double quotes around it, even if the element was a string.

Fixed a bug where an extra set of double quotes was being added at the
beginning and end of each csv row string.

Cleaned up logic to generate a csv row string from an array of elements.