Run background service worker as ES Module
background.js is now loaded directly instead of imported through
background-worker.js. Since the latter is no longer necessary, it
should be removed shortly.
There are also other files that are/were also only still in use because
service workers couldn't be loaded as ES modules, and these should be
deprecated in the near future as well.
The functionality related to this change became available as of
Chrome versions ~91-93, when Chromium bug 824647 and others
were addressed.