20221004 modularize background serviceworker (#25)
Load background script as an ES Module:
- background.js is now loaded directly instead of imported through
background-worker.js. Since the latter is no longer necessary, it
should be removed in a future update.
- There are also other files that are/were also only still in use because
service workers couldn't be loaded as ES modules, and these have 
now been removed.
- The functionality related to this change became available as of
Chrome versions ~91-93, when Chromium bug 824647 and others 
were addressed.

Miscellaneous:
- Corrected bug number blocking Firebase Auth UI. It was listed 
as 1247690, but 1255412 is more relevant. Note that this is a 
Chromium bug, and is unlikely to get fixed. 
- Removed capitalization from config file