Leverage scraped metadata to enable extension icon on page refresh (#26)
The 'header' element in the DOM is now scraped when the content script
is first loaded, so that tracklist metadata can be extracted. This
means the extension icon will get updated properly if the page is
refreshed.

Pulled the logic to scrape tracklist metadata from the DOM 'header'
element into a dedicated function for better readability & reusability.

Replaced the overly convoluted getMetadatum function with two simpler
functions to scrape the tracklist track count and title, respectively.