Moved some GPM related logic into gpm-storage.js
Moved the logic to get a list of all gpm tracklist titles and the track
count of a specific tracklist out of the background file and into
gpm-storage.js.
Copied the logic to get the tracks array for a specific tracklist from
StorageManager.js into gpm-storage.js. This logic still needs to be
removed from the former, and any references updated.
gpm-storage.js now saves the GPM Library data object in a constant
session state variable so that it doesn't have to be fetched from
Chrome storage more than once per session.