Add getGpmData fnc to handle GPM related requests
The new function handles various different requests and consolidates
logic for similar requests that used to leverage duplicated / reduntant
code and calls to Chrome local storage.
The GPM Library data is now fetched from Chrome local storage the first
time it's requested and then saved in session state for any future
requests.
The data is also saved in session state in a format that more closely
matches how YTM data is stored, so the data can be accessed more easily.

The list of Uploaded GPM tracks is also saved in Chrome local storage,
so when it's requested it's fetched from there instead of calculated
from scratch. It's only re-calculated as a fall-back.

Note this work is still in progress. Not all functionality has been
ported over to the new function yet, and its structure may still change.