Save gpm uploads as tracklist in session state
The list of GPM Uploaded tracks is now saved as a tracklist under the
'Uploaded Songs' key in the gpmTracklists object of sesstion state,
alongside all the other tracklists.
Previously, it was saved as a tracks array separately from all the
other tracklists. Making it more consistent with the rest of the
tracklists will require fewer extra steps later on and easier to read.