Move check for 'Uploaded Songs'
A special check is needed when a request is received to return the GPM
list of 'Uploaded Songs'. Previously this was handed in the request to
get the 'tracksArray'. It's now been moved to the request to get the
'tracklist', which is a bit more logical, as the tracklist itself
should be the first point of entry for any gpm tracklist related
request.