Get 'tracklists' & 'tracksNotInCommon' via new fnc
Added a getGpmLibraryData function to handle cases where the requested
data is not specific to one particular tracklist, but rather the
broader GPM library. This includes the following requests:
'tracklists', 'allPlaylists', 'tracklistTitles', 'tracksNotInCommon'.
Note that 'allPlaylists' & 'tracklistTitles' are not yet being accessed
through this function, but that will be addressed in a later change.