Rename getGpmData
The original function getGpmData has been split into two:
- getTracklistData() to handle requests related to a specific tracklist
- getLibraryData() to handle requests related to the broaded GPM library

The logic has now been properly split between them and the calls
throughout the rest of the app have been updated.

Also, a raw copy of the GPM library data, as it's strored in Chrome
local storage, is no longer saved in session state, because it was not
being used. The list of generated tracklist data objects is used
instead.