Consolidated GPM-related logic into its own file (#28)
Moved all logic related to retrieving GPM library data to its own file
in a new 'storage' directory.  This consolidates logic that was
previously duplicated and scattered across multiple files into one
place, which makes it much easier to read and follow.

The new gpm-storage.js file will fetch GPM library data from Chrome
local storage only a single time, whereas before this was done
repeatedly, on demand. Now, the first time any GPM data is requested,
it is fetched from local storage, and a list of tracklist data objects
is then generated from the raw list of GPM tracks arrays. These
tracklist data objects are then stored in a constant session state
variable in a format that more closely matches how YTM data is
stored. This makes any subsequent requests for GPM related data much
simpler and quicker.

The new file has a getLibraryData function which supports various
requests, including fetching multiple tracklists, a list of all
'playlists', and a list of all tracklist titles.

This file also has a getTracklistData function that supports requests
to return a tracklist, a tracks array, or a track count. This function
also now directly handles the special case of the list of uploaded
songs, which requires special steps to retrieve or generate. Again,
this code was previously duplicated across multiple files and functions
and is now much more concise.

Note that the list of Uploaded GPM tracks is also saved in Chrome local
storage, so when it's first requested it's fetched from there instead
of generated from scratch. It's only regenerated as a fall-back. The
first time the list is retrieved or generated, it is saved as a
tracklist data object along with the rest of the tracklists in session
state. This makes subsequent requests for this list much quicker and
simpler, and more consistent with all the other tracklists.

StorageManager.js was moved to the new storage directory and renamed to
firestore-storage.js to more accurately reflect its purpose, since
requests to get GPM library date in Chrome local storage have now been
moved out of this file. There are still some remaining requests related
to data in Chrome sync storage, but these will be moved in a future
update. Ultimately, this file should be dedicated to accessing data
contained in the Firebase Firestore database (i.e YTM data).