Consolidate options storage logic & detach it from options UI code (#30)
Changes to Options Code:
Split code related to extension options into two files: 
options-storage.js & options-ui.js. They have also both been moved into 
a new dedicated directory for options logic. 
The options-storage file handles any calls to Chrome storage as needed, 
and options-ui only interacts directly with options-storage. 

Previoulsy options.js had its own logic to fetch the preferences
storage item from Chrome sync storage and then update the comparison
method property manually. Now, it simply uses the newly available
modifyStorageItemProperty helper function (described below).

The options-storage file now uses getValueAtKey instead of 
getKeyValuePairs when requesting preferences from storage, which is a 
bit simpler and more appropriate in this case. 

The Storage Item Key and Storage Item Properties are now defined in a 
single location in the options-storage file. This is intened to reduce 
the chances of user error and to increase readability. A similar 
pattern may be used in other files in the future, if deemed helpful. 

Updated event-controller to leverage the new options-storage file to 
get preferences, and removed the previous (now obsolete) functionality 
from the firestore-storage file. 

Utilities:
Added a helper function to modify the property of an item stored in 
Chrome storage. When this needs to be done, the entire storage item 
has to be fetched, updated, and then saved back to storage. Now the 
code to handle this has been consolidated to one helper, rather than 
needing to be duplicated across various files. 
Note that if the storage item to be modified doesn't exist, a new empty 
object will be created with the desired property and then saved to 
storage. 
Note that some files still need to be updated to use this new helper. 

Updated chrome-storage-promises to use more consistent terminology, 
and fixed some of the JSDoc documentation. 

Bug Fixes:
Fixed a bug when saving GPM tracklists to session state. After merging 
two functions in the previous update, an incorrect variable was being 
used when saving the GPM tracklists to session state.

Miscellaneous:
Removed an obsolete function from event-controller that should have 
been removed in the previous update.