Add ChromeStorageAccessor class & new generic storage file (#36)
ChromeStorageAccessor Class:
The new ChromeStorageAccessor class is in a new chrome-storage.js file. 
Added various tests for this class in a new chrome-storage-tests.js 
file. The tests are written such that they can be executed in series. 

The new class has support for easily getting or updating a storage 
item's property value, and removes the need for duplicated code across 
the codebase. 

The new class includes an improved check for a storage item having a 
given property. Checking if the property is of type 'undefined' isn't 
ideal for storage items, since they won't store undefined properties 
anyway. Checking if the property actually exists makes more sense.

The class includes functionality to set the storage item key after the 
class has been instantiated. This is because, currently, the storage 
item key is not always known when the accessor is created. 

The function to set a property's value includes an option to not 
override the current value, if it exists. This makes it easy to set a 
default property value. 

Note that setting the property of a chrome storage item to undefined is 
currently allowed, and this will result in the property being removed 
from the storage item, because of the way Chrome handles storage items. 

New Storage File:
Added storage.js to control all (or most) storage interactions. 

Storage.js will handle all or most high-level interactions related
to storage. The functions here will leverage the more specific storage
files/mechanisms (e.g. firestore, chrome storage, etc.) as needed,
while abstracting these details away from other files like
event-controller.js, which shouldn't necesseraily track where different
data is stored, but rather simply request to retrieve or store the data 
as needed. 

Currently, some Firestore data is still accessed directly, but this 
may be modified in a future update. 

Various logic that is now consolidated in storage.js was previously in 
background.js and firestore-storage.js. Now that it is all in one 
location, it should be easier to read and follow. 

Miscellaneous:
Removed incorrect checks for runtime.lastError. 
Removed unnecessary extra checks when logging out. Previously, when 
logging out, background.js would access the current tracklist metadata 
before updating the extension icon. This is unnecessary because the 
metadata isn't needed to simply change the icon text to read "login".