Add session-state.js to hold session data
Previously, any lists of tracks or other data that has either been
scraped from the DOM or fetched from storage would be tracked in
event-controller.js itself. Pulling this logic into a separate file is
intended to make it easier to follow the flow of the code, and to allow
event-controller.js to be focused more on reacting and responding to
events, rather than passing around pieces of data between other files.

This is still a work in progress and needs considerable more work
before it gets to that point. The first step done here is to create the
new session-state.js file and to set up the interactions between that
and event-controller.js. Many of these may be modified in future
changes.