Move logic to download csv into separate file
Previously this logic was being handled in event-controller.js. It is
now handled in a new export-tracklists.js, in a new io directory. For
better readability, the helper function has been split into three, one
each for scraped, stored, and gpm tracks.

Note that error checking was removed and needs to be put back in.